ai 翻譯機器人/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==0.27.8
